+++ tests/_test.py
@@ -15,6 +15,7 @@ import docsig.messages
 
 from . import (
     CHECK_CLASS,
+    CHECK_PROTECTED,
     E10,
     ERR_GROUP,
     FUNC,
@@ -69,7 +70,9 @@ def test_main_args(
     :param template: Contents to write to file.
     """
     file = init_file(template)
-    assert main(CHECK_CLASS, file.parent) == int(name.startswith("fail"))
+    assert main(CHECK_CLASS, CHECK_PROTECTED, file.parent) == int(
+        name.startswith("fail")
+    )
 
 
 @pytest.mark.parametrize(
@@ -95,7 +98,7 @@ def test_main_output(
     :param template: String data.
     """
     file = init_file(template.template)
-    main(CHECK_CLASS, file.parent)
+    main(CHECK_CLASS, CHECK_PROTECTED, file.parent)
     assert template.expected in nocolorcapsys.readouterr()[0]
 
 
@@ -380,7 +383,7 @@ def test_main_str(
     :param name: Name of test.
     :param template: Contents to write to file.
     """
-    assert main(CHECK_CLASS, "--string", template) == int(
+    assert main(CHECK_CLASS, CHECK_PROTECTED, "--string", template) == int(
         name.startswith("fail")
     )
 
@@ -405,7 +408,7 @@ def test_main_str_out(
         color codes.
     :param template: String data.
     """
-    main(CHECK_CLASS, "--string", template.template)
+    main(CHECK_CLASS, CHECK_PROTECTED, "--string", template.template)
     assert template.expected in nocolorcapsys.readouterr()[0]
 
 
@@ -423,5 +426,31 @@ def test_no_check_init_flag(
     """
     template = templates.registered.getbyname("fail-init")
     file = init_file(template.template)
+    assert main(file.parent) == 0
+    assert not nocolorcapsys.stdout()
+
+
+@pytest.mark.parametrize(
+    ["_", TEMPLATE, "__"],
+    templates.registered.getgroup("fail-protect"),
+    ids=templates.registered.getgroup("fail-protect").getids(),
+)
+def test_no_check_protected_flag(
+    init_file: InitFileFixtureType,
+    main: MockMainType,
+    nocolorcapsys: NoColorCapsys,
+    _: str,
+    template: str,
+    __: str,
+) -> None:
+    """Test that failing func passes without ``--check-protected`` flag.
+
+    :param init_file: Initialize a test file.
+    :param main: Mock ``main`` function.
+    :param nocolorcapsys: Capture system output while stripping ANSI
+        color codes.
+    :param template: Contents to write to file.
+    """
+    file = init_file(template)
     assert main(file.parent) == 0
     assert not nocolorcapsys.stdout()